# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-checklist
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /employee-checklists/{employeeChecklistId}/phases/{phaseId}/customtasks:
    post:
      tags:
      - Employee checklist resources
      summary: Create a custom task
      description: Create a custom task connected to a specific employee checklist
      operationId: createCustomTask
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomTaskCreateRequest"
        required: true
      responses:
        "201":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomTask"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/CustomTask"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /employee-checklists/initialize:
    post:
      tags:
      - Employee checklist resources
      summary: Inititalize checklists for new employees
      description: Trigger creation of checklists for all known new employees
      operationId: initiateEmployeeChecklists
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistResponse"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /employee-checklists/initialize/{personId}:
    post:
      tags:
      - Employee checklist resources
      summary: Inititalize checklists for a specific employee
      description: Trigger creation of checklist for employee matching sent in person
        id
      operationId: initiateSpecificEmployeeChecklist
      parameters:
      - name: personId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistResponse"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /employee-checklists/{employeeChecklistId}/tasks/{taskId}:
    patch:
      tags:
      - Employee checklist resources
      summary: Update fulfilment of a task
      operationId: updateTaskFulfilment
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeChecklistTaskUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistTask"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistTask"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /employee-checklists/{employeeChecklistId}/phases/{phaseId}:
    patch:
      tags:
      - Employee checklist resources
      summary: Update of all tasks in a phase
      description: Bulk update of sent in attributes for all tasks in phase
      operationId: updateAllTasksInPhase
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeChecklistPhaseUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPhase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPhase"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /employee-checklists/{employeeChecklistId}/customtasks/{taskId}:
    get:
      tags:
      - Employee checklist resources
      summary: Read a custom task
      description: Read a custom task connected to a specific employee checklist
      operationId: readCustomTask
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomTask"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/CustomTask"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Employee checklist resources
      summary: Delete a custom task
      description: Delete a custom task from an employee checklist
      operationId: deleteCustomTask
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Employee checklist resources
      summary: Update a custom task
      description: Update a custom task connected to a specific employee checklist
      operationId: updateCustomTask
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomTaskUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomTask"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/CustomTask"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /checklists/{checklistId}:
    get:
      tags:
      - Checklist resources
      summary: Fetch checklist by id
      operationId: fetchChecklistById
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Checklist resources
      summary: Delete a checklist
      operationId: deleteChecklist
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Checklist resources
      summary: Update a checklist
      operationId: updateChecklist
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /checklists/{checklistId}/phases/{phaseId}:
    get:
      tags:
      - Phase resources
      summary: Fetch phase in a checklist
      operationId: fetchChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Phase"
    delete:
      tags:
      - Phase resources
      summary: Delete phase in a checklist
      operationId: deleteChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Successful Operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Phase resources
      summary: Update phase in a checklist
      operationId: updateChecklistPhase
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhaseUpdateRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Phase"
...
  /employee-checklists/search:
    get:
      tags:
      - Employee checklist resources
      summary: |
        Fetches employee checklists based on search string, the string is matched against the following
        attributes using 'LikeIgnoreCase':
        - Employee's first name
        - Employee's last name
        - Employee's username
        - Employee's company name
        - Employee's manager first name
        - Employee's manager last name
        - Delegate's first name
        - Delegate's last name
      operationId: findemployeeChecklistsBySearchString
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: searchString
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPaginatedResponse"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistPaginatedResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
...
    CustomTaskCreateRequest:
      required:
      - heading
      - questionType
      - sortOrder
      type: object
      properties:
        heading:
          type: string
          description: The heading of the task
          writeOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          writeOnly: true
          example: Detta är en beskrivning av ett uppdrag
        questionType:
          $ref: "#/components/schemas/QuestionType"
        sortOrder:
          type: integer
          description: The sort order for the task
          format: int32
          writeOnly: true
      description: Model for custom task create request
...
    CustomTask:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order for the task
          format: int32
          readOnly: true
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        created:
          type: string
          description: The date and time the task was created
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Model for custom task
...
    EmployeeChecklistPhaseUpdateRequest:
      type: object
      properties:
        tasksFulfilmentStatus:
          $ref: "#/components/schemas/FulfilmentStatus"
      description: Model for update request on an employee checklist phase
...
  securitySchemes: {}